requirements.txt
psycopg2-binary = ">=2.7.5"
django-widget-tweaks = ">=1.4.2"
django-cleanup = ">=2.1.0"
django-countries = ">=5.3.2"
django-user-language-middleware = ">=0.0.2"
django-currentuser = ">=0.2.3"
django-tagulous = ">=0.13.2"
django-tables2 = ">=2.0.2"
django-storages = ">=1.7.1"
django-debug-toolbar = ">=1.9.1"
ipdb = ">=0.11"
Django = "==2.0.8"
Pillow = ">=5.2.0"
Unidecode = ">=1.0.22"
rabbitmq = "*"
pika = "*"
paramiko = "*"
django-minio-storage = "*"
djangorestframework = "*"
markdown = "*"
django-filter = "*"
numpy = "*"

